styled-component _document.tsx issue fix, search component 분리 작업 진행 중
