+++ src/styles/search/detailStyle.tsx
@@ -1,102 +1,5 @@
 import styled from "styled-components";
 
-const SearchInit = styled.div`
-  text-align: center;
-  font-size: 14px;
-  line-height: 20px;
-`;
-const CategorySubWrapper = styled.div`
-  position: relative;
-  display: flex;
-  flex-wrap: wrap;
-`;
-const CategoryListWrapper = styled.ul`
-  margin: 10px 10px 10px 20px;
-`;
-
-const LiLink = styled.li`
-  float: left;
-  margin: 0 10px 10px 0;
-  cursor: pointer;
-  display: block;
-`;
-
-const BtnTag = styled.div`
-  color: #677ac7;
-  font-size: 14px;
-  line-height: 32px;
-  padding: 0 10px 0 10px;
-  display: inline-block;
-  border: 1px solid #dddddd;
-  box-sizing: border-box;
-  border-radius: 10px;
-  border: 1px solid #677ac7;
-  &.on {
-    background: #677ac7;
-    color: #ffffff;
-  }
-`;
-const TabSubWrapper = styled.div`
-  position: relative;
-`;
-const NotFoundWrapper = styled.div`
-  position: relative;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-`;
-const Svgbox = styled.div`
-  position: relative;
-  padding: 109px 0 0 0;
-  font-weight: 500;
-  font-size: 20px;
-`;
-const NotFoundMain = styled.div`
-  position: relative;
-  padding: 32px 0 0 0;
-  font-weight: 500;
-  font-size: 20px;
-`;
-const NotFoundSub = styled.div`
-  position: relative;
-  padding: 10px 0 109px 0;
-  line-height: 20px;
-  color: #222222;
-  font-size: 14px;
-`;
-const Ul = styled.ul`
-  display: flex;
-  height: 80px;
-`;
-const UlDiv = styled.div`
-  flex: 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  transition: all 0.3s;
-  cursor: pointer;
-  font-size: 14px;
-  line-height: 125.69%;
-  color: #999999;
-  border-bottom: 2px solid #f1f1f3;
-  &.on {
-    font-weight: 500;
-    color: #222222;
-    border-bottom: 2px solid #ffd262;
-  }
-`;
-
-const FilterInnerWrapper = styled.div`
-  position: relative;
-  margin: 10px 20px 10px;
-`;
-const FilterTitle = styled.div`
-  font-family: "Nunito", sans-serif;
-  font-size: 14px;
-  line-height: 40px;
-  color: #777777;
-`;
 const ListHeadWrapper = styled.div`
   position: relative;
   margin: 10px 20px 10px;
@@ -316,20 +219,6 @@ const ListFooterIconBox = styled.div`
   margin: 0 2px;
 `;
 export {
-  SearchInit,
-  CategorySubWrapper,
-  CategoryListWrapper,
-  LiLink,
-  BtnTag,
-  TabSubWrapper,
-  NotFoundWrapper,
-  Svgbox,
-  NotFoundMain,
-  NotFoundSub,
-  Ul,
-  UlDiv,
-  FilterInnerWrapper,
-  FilterTitle,
   ListHeadWrapper,
   ListHeadTitle,
   ListHeadDetail,
